Refactor make dataset card tests
