--- tests/test_card.py
@@ -1,2064 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-import random
-import unittest
-from pathlib import Path
-
-import sys
-
-PROJECT_ROOT = Path(__file__).resolve().parents[1]
-SRC_PATH = PROJECT_ROOT / "src"
-if str(SRC_PATH) not in sys.path:
-    sys.path.insert(0, str(SRC_PATH))
-
-import numpy as np
-from ase import Atoms
-from ase.io import read
-from PySide6.QtWidgets import QApplication
-
-from NepTrainKit.ui.views._card import (
-    SuperCellCard,
-    CrystalPrototypeBuilderCard,
-    PerturbCard,
-    CellScalingCard,
-    CellStrainCard,
-    ShearMatrixCard,
-    ShearAngleCard,
-    RandomSlabCard,
-    RandomDopingCard,
-    CompositionSweepCard,
-    CompositionGradientCard,
-    RandomOccupancyCard,
-    ConditionalReplaceCard,
-    MagneticOrderCard,
-    SetMagneticMomentsCard,
-    SmallAngleSpinTiltCard,
-    SpinDisorderCard,
-    SpinSpiralCard,
-    FoldedHelixCard,
-    MagneticMomentRotationCard,
-    GroupLabelCard,
-    RandomVacancyCard,
-    VacancyDefectCard,
-    StackingFaultCard,
-    InsertDefectCard,
-    LayerCopyCard,
-    OrganicMolConfigPBCCard,
-    VibrationModePerturbCard,
-    GeometryFilterCard,
-)
-from NepTrainKit.core.magnetism import orthonormal_frame
-from NepTrainKit.core import CardManager
-from NepTrainKit.core.alloy import parse_composition
-from NepTrainKit.core.cards.alloy import (
-    CompositionSweepOperation,
-    CompositionSweepParams,
-    CompositionGradientOperation,
-    CompositionGradientParams,
-    ConditionalReplaceOperation,
-    ConditionalReplaceParams,
-    RandomDopingOperation,
-    RandomDopingParams,
-    RandomOccupancyOperation,
-    RandomOccupancyParams,
-)
-from NepTrainKit.core.cards.filter import FPSFilterOperation, FPSFilterParams, GeometryFilterOperation, GeometryFilterParams
-from NepTrainKit.core.cards.operation import params_to_dict
-from NepTrainKit.core.cards.structure import (
-    CrystalPrototypeBuilderOperation,
-    CrystalPrototypeBuilderParams,
-    GroupLabelOperation,
-    GroupLabelParams,
-    LayerCopyOperation,
-    LayerCopyParams,
-    OrganicMolConfigPBCParams,
-    VibrationModePerturbOperation,
-    VibrationModePerturbParams,
-)
-from NepTrainKit.core.cards.defect import (
-    InsertDefectOperation,
-    InsertDefectParams,
-    RandomSlabOperation,
-    RandomSlabParams,
-    RandomVacancyOperation,
-    RandomVacancyParams,
-    StackingFaultOperation,
-    StackingFaultParams,
-    VacancyDefectOperation,
-    VacancyDefectParams,
-)
-from NepTrainKit.core.cards.lattice import (
-    CellScalingOperation,
-    CellScalingParams,
-    CellStrainOperation,
-    CellStrainParams,
-    PerturbOperation,
-    PerturbParams,
-    ShearAngleOperation,
-    ShearAngleParams,
-    ShearMatrixOperation,
-    ShearMatrixParams,
-    SuperCellOperation,
-    SuperCellParams,
-)
-from NepTrainKit.core.cards.magnetism import (
-    FoldedHelixOperation,
-    FoldedHelixParams,
-    MagneticMomentRotationOperation,
-    MagneticMomentRotationParams,
-    MagneticOrderOperation,
-    MagneticOrderParams,
-    SetMagneticMomentsOperation,
-    SetMagneticMomentsParams,
-    SmallAngleSpinTiltOperation,
-    SmallAngleSpinTiltParams,
-    SpinDisorderOperation,
-    SpinDisorderParams,
-    SpinSpiralOperation,
-    SpinSpiralParams,
-)
-from NepTrainKit.ui.widgets import MakeDataCard
-from NepTrainKit.ui.widgets.card_metadata import card_tooltip, metadata_html
-from NepTrainKit.ui.widgets.doping_rule import DopingRuleItem
-from NepTrainKit.ui.widgets.vacancy_rule import VacancyRuleItem
-from NepTrainKit.version import DOCS_BASE_URL
-
-BASE_CARD_KEYS = {"class", "check_state", "metadata", "params"}
-
-
-class _ExternalTestCard(MakeDataCard):
-    card_name = "External Test Card"
-
-    def process_structure(self, structure):
-        return [structure]
-
-
-@CardManager.register_card
-class _MetadataTestCard(MakeDataCard):
-    card_name = "Metadata Test Card"
-    description = "Card used to verify contributor metadata."
-    card_version = "0.1"
-    contributors = [
-        {
-            "name": "Test Contributor",
-            "role": "author",
-            "email": "test@example.com",
-            "url": "https://example.com/test",
-            "affiliation": "Test Lab",
-        }
-    ]
-
-    def process_structure(self, structure):
-        return [structure]
-
-
-class TestCard(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        cls._app = QApplication.instance() or QApplication([])
-        cls.test_dir = Path(__file__).parent
-        cls.base_structure = read(cls.test_dir / "data" / "Si2.vasp")
-        cls.base_structure.info.setdefault("Config_type", "Si2")
-
-    @classmethod
-    def tearDownClass(cls):
-        if cls._app is not None:
-            cls._app.quit()
-            cls._app = None
-
-    def setUp(self):
-        np.random.seed(0)
-        random.seed(0)
-        self.structure = self.base_structure.copy()
-
-    def test_builtin_card_has_online_doc_url(self):
-        card = StackingFaultCard()
-
-        self.assertEqual(
-            card.get_online_doc_url(),
-            f"{DOCS_BASE_URL}module/make-dataset-cards/cards/stacking-fault-card.html",
-        )
-        self.assertFalse(card.doc_button.isHidden())
-
-    def test_external_card_hides_online_doc_button(self):
-        card = _ExternalTestCard()
-
-        self.assertEqual(card.get_online_doc_url(), "")
-        self.assertTrue(card.doc_button.isHidden())
-
-    def test_card_contributor_metadata_includes_optional_email(self):
-        metadata = CardManager.get_card_metadata("_MetadataTestCard")
-
-        self.assertIsNotNone(metadata)
-        self.assertEqual(metadata.card_name, "Metadata Test Card")
-        self.assertEqual(metadata.version, "0.1")
-        self.assertEqual(metadata.contributors[0].email, "test@example.com")
-        self.assertIn("Test Contributor", card_tooltip(metadata))
-        self.assertIn("mailto:test@example.com", metadata_html(metadata))
-
-        card = _MetadataTestCard()
-        data = card.to_dict()
-        self.assertEqual(data["metadata"]["contributors"], ["Test Contributor"])
-        self.assertEqual(data["metadata"]["card_version"], "0.1")
-
-    def test_operation_cards_write_only_params(self):
-        for class_name, card_cls in CardManager.card_info_dict.items():
-            if not hasattr(card_cls, "create_operation"):
-                continue
-            card = card_cls()
-            if card.create_operation() is None:
-                continue
-            serialized = card.to_dict()
-            self.assertEqual(
-                set(serialized),
-                BASE_CARD_KEYS,
-                f"{class_name} should write only current params format",
-            )
-
-    def test_legacy_card_keys_still_load(self):
-        strain = CellStrainCard()
-        strain.from_dict(
-            {
-                "check_state": True,
-                "organic": True,
-                "engine_type": "biaxial",
-                "x_range": [1.0, 2.0, 0.5],
-                "y_range": [3.0, 4.0, 0.5],
-                "z_range": [0.0, 0.0, 1.0],
-            }
-        )
-        self.assertEqual(
-            strain.get_params(),
-            CellStrainParams(
-                axes="biaxial",
-                x_range=(1.0, 2.0, 0.5),
-                y_range=(3.0, 4.0, 0.5),
-                z_range=(0.0, 0.0, 1.0),
-                identify_organic=True,
-            ),
-        )
-
-        layer = LayerCopyCard()
-        layer.from_dict(
-            {
-                "check_state": True,
-                "preset_index": 0,
-                "dz_expr": "A + z*0",
-                "params": "A=1.5",
-                "apply_mode": 2,
-                "elements": "Si",
-                "z_range": [0.0, 2.0],
-                "wrap": True,
-                "extend_cell_z": False,
-                "extra_vacuum": [1.0],
-                "layers": [2],
-                "distance": [4.0],
-            }
-        )
-        self.assertEqual(
-            layer.get_params(),
-            LayerCopyParams(
-                preset_index=0,
-                dz_expr="A + z*0",
-                expression_params="A=1.5",
-                apply_mode=2,
-                elements="Si",
-                z_range=(0.0, 2.0),
-                wrap=True,
-                extend_cell_z=False,
-                extra_vacuum=1.0,
-                layers=2,
-                distance=4.0,
-            ),
-        )
-
-        operation_params = LayerCopyParams(
-            preset_index=1,
-            dz_expr="z + 1",
-            expression_params="",
-            apply_mode=1,
-            elements="",
-            z_range=(-1.0, 1.0),
-            wrap=False,
-            extend_cell_z=True,
-            extra_vacuum=0.5,
-            layers=4,
-            distance=2.5,
-        )
-        layer.from_dict(
-            {
-                "check_state": True,
-                "operation_params": params_to_dict(operation_params),
-            }
-        )
-        self.assertEqual(layer.get_params(), operation_params)
-
-    @staticmethod
-    def _spin_chain():
-        atoms = Atoms(
-            symbols=["Fe", "Fe", "Fe", "Fe"],
-            positions=[
-                [0.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0],
-                [0.0, 0.0, 2.0],
-                [0.0, 0.0, 3.0],
-            ],
-            cell=np.diag([6.0, 6.0, 8.0]),
-            pbc=[False, False, True],
-        )
-        atoms.info["Config_type"] = "Fe_chain"
-        return atoms
-
-    @staticmethod
-    def _spin_bilayer_chain():
-        atoms = Atoms(
-            symbols=["Fe"] * 8,
-            positions=[
-                [0.0, 0.0, 0.00],
-                [1.0, 0.0, 0.01],
-                [0.0, 0.0, 1.00],
-                [1.0, 0.0, 1.01],
-                [0.0, 0.0, 2.00],
-                [1.0, 0.0, 2.01],
-                [0.0, 0.0, 3.00],
-                [1.0, 0.0, 3.01],
-            ],
-            cell=np.diag([4.0, 4.0, 6.0]),
-            pbc=[False, False, True],
-        )
-        atoms.info["Config_type"] = "Fe_bilayer_chain"
-        return atoms
-
-    def test_supercell_card_variants(self):
-        card = SuperCellCard()
-        structure = self.structure.copy()
-        original_cell = np.array(structure.get_cell())
-
-        card.super_scale_condition_frame.set_input_value([2, 1, 1])
-        card.super_scale_radio_button.setChecked(True)
-        card.super_cell_radio_button.setChecked(False)
-        card.max_atoms_radio_button.setChecked(False)
-        card.behavior_type_combo.setCurrentIndex(0)
-
-        direct_results = card.process_structure(structure)
-        self.assertEqual(len(direct_results), 1)
-        self.assertEqual(len(direct_results[0]), len(structure) * 2)
-        new_cell = np.array(direct_results[0].get_cell())
-        self.assertGreater(np.linalg.norm(new_cell[0]), np.linalg.norm(original_cell[0]))
-
-        card.super_cell_radio_button.setChecked(True)
-        card.super_scale_radio_button.setChecked(False)
-        card.max_atoms_radio_button.setChecked(False)
-        lengths = structure.cell.lengths()
-        card.super_cell_condition_frame.set_input_value([
-            lengths[0] * 2.1,
-            lengths[1] * 1.1,
-            lengths[2] * 1.1,
-        ])
-
-        cell_results = card.process_structure(structure)
-        self.assertEqual(len(cell_results), 1)
-        self.assertGreater(len(cell_results[0]), len(structure))
-
-        card.behavior_type_combo.setCurrentIndex(1)
-        card.max_atoms_radio_button.setChecked(True)
-        card.super_cell_radio_button.setChecked(False)
-        card.super_scale_radio_button.setChecked(False)
-        card.max_atoms_condition_frame.set_input_value([len(structure) * 2])
-
-        atoms_results = card.process_structure(structure)
-        self.assertGreaterEqual(len(atoms_results), 1)
-        self.assertTrue(any(len(atoms) > len(structure) for atoms in atoms_results))
-
-    def test_supercell_card_fixed_axis_lock(self):
-        card = SuperCellCard()
-        structure = self.structure.copy()
-        original_cell = np.array(structure.get_cell())
-        original_lengths = np.linalg.norm(original_cell, axis=1)
-
-        card.behavior_type_combo.setCurrentIndex(1)
-        card.super_cell_radio_button.setChecked(True)
-        card.super_scale_radio_button.setChecked(False)
-        card.max_atoms_radio_button.setChecked(False)
-        card.fixed_axis_c_checkbox.setChecked(True)
-        card.fixed_scale_condition_frame.set_input_value([1, 1, 1])
-        card.super_cell_condition_frame.set_input_value([
-            original_lengths[0] * 2.1,
-            original_lengths[1] * 2.1,
-            original_lengths[2] * 4.0,
-        ])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 4)
-        self.assertTrue(
-            all(
-                np.isclose(np.linalg.norm(np.array(atoms.get_cell())[2]), original_lengths[2], atol=1e-6)
-                for atoms in results
-            )
-        )
-        self.assertTrue(
-            any(
-                np.linalg.norm(np.array(atoms.get_cell())[0]) > original_lengths[0] + 1e-6
-                or np.linalg.norm(np.array(atoms.get_cell())[1]) > original_lengths[1] + 1e-6
-                for atoms in results
-            )
-        )
-
-        data = card.to_dict()
-        restored = SuperCellCard()
-        restored.from_dict(data)
-        self.assertTrue(restored.fixed_axis_c_checkbox.isChecked())
-        self.assertEqual(restored.fixed_scale_condition_frame.get_input_value(), [1, 1, 1])
-        self.assertEqual(restored.super_cell_condition_frame.get_input_value(), list(card.super_cell_condition_frame.get_input_value()))
-
-    def test_supercell_operation_matches_card_params(self):
-        card = SuperCellCard()
-        structure = self.structure.copy()
-        card.super_scale_radio_button.setChecked(True)
-        card.super_cell_radio_button.setChecked(False)
-        card.max_atoms_radio_button.setChecked(False)
-        card.super_scale_condition_frame.set_input_value([2, 1, 1])
-
-        params = card.get_params()
-        self.assertIsInstance(params, SuperCellParams)
-        card_result = card.process_structure(structure)
-        op_result = SuperCellOperation().run_structure(structure, params)
-
-        self.assertEqual(len(card_result), len(op_result))
-        self.assertEqual(len(card_result[0]), len(op_result[0]))
-        self.assertIn("params", card.to_dict())
-
-    def test_perturb_card_with_organic(self):
-        card = PerturbCard()
-        structure = self.structure.copy()
-        card.organic_checkbox.setChecked(True)
-        card.engine_type_combo.setCurrentIndex(1)
-        card.scaling_condition_frame.set_input_value([0.1])
-        card.num_condition_frame.set_input_value([2])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        displacements = [
-            np.linalg.norm(atoms.get_positions() - structure.get_positions(), axis=1).max()
-            for atoms in results
-        ]
-        self.assertTrue(any(delta > 0 for delta in displacements))
-
-    def test_perturb_operation_is_ui_independent(self):
-        params = PerturbParams(
-            engine_type=1,
-            max_distance=0.1,
-            max_num=2,
-            use_seed=True,
-            seed=11,
-        )
-        results = PerturbOperation().run_structure(self.structure.copy(), params)
-
-        self.assertEqual(len(results), 2)
-        self.assertTrue(all("Pert(d=0.1,U)" in atoms.info.get("Config_type", "") for atoms in results))
-
-    def test_cell_scaling_card_options(self):
-        card = CellScalingCard()
-        structure = self.structure.copy()
-        card.organic_checkbox.setChecked(True)
-        card.perturb_angle_checkbox.setChecked(False)
-        card.engine_type_combo.setCurrentIndex(1)
-        card.scaling_condition_frame.set_input_value([0.05])
-        card.num_condition_frame.set_input_value([2])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        original_lengths = np.array(structure.cell.cellpar()[:3])
-        self.assertTrue(
-            any(
-                not np.allclose(
-                    np.array(atoms.cell.cellpar()[:3]),
-                    original_lengths,
-                    atol=1e-6,
-                )
-                for atoms in results
-            )
-        )
-
-    def test_cell_scaling_operation_is_ui_independent(self):
-        params = CellScalingParams(
-            engine_type=1,
-            max_scaling=0.05,
-            max_num=2,
-            perturb_angle=False,
-            use_seed=True,
-            seed=7,
-        )
-        results = CellScalingOperation().run_structure(self.structure.copy(), params)
-
-        self.assertEqual(len(results), 2)
-        self.assertTrue(all("LSc(max=0.05,U)" in atoms.info.get("Config_type", "") for atoms in results))
-
-    def test_cell_strain_card_uniaxial(self):
-        card = CellStrainCard()
-        structure = self.structure.copy()
-        card.organic_checkbox.setChecked(True)
-        card.engine_type_combo.setText("uniaxial")
-        card.strain_x_frame.set_input_value([1.0, 1.0, 1.0])
-        card.strain_y_frame.set_input_value([0.0, 0.0, 1.0])
-        card.strain_z_frame.set_input_value([0.0, 0.0, 1.0])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 3)
-        original_lengths = np.array(structure.cell.cellpar()[:3])
-        self.assertTrue(
-            any(
-                not np.allclose(
-                    np.array(atoms.cell.cellpar()[:3]),
-                    original_lengths,
-                    atol=1e-6,
-                )
-                for atoms in results
-            )
-        )
-
-    def test_cell_strain_operation_is_ui_independent(self):
-        params = CellStrainParams(
-            axes="X",
-            x_range=(1.0, 1.0, 1.0),
-            y_range=(0.0, 0.0, 1.0),
-            z_range=(0.0, 0.0, 1.0),
-            identify_organic=False,
-        )
-        results = CellStrainOperation().run_structure(self.structure.copy(), params)
-
-        self.assertEqual(len(results), 1)
-        self.assertIn("Str(X=1%)", results[0].info.get("Config_type", ""))
-
-    def test_fps_filter_operation_rejects_missing_model(self):
-        params = FPSFilterParams(nep_path=str(self.test_dir / "data" / "missing_nep.txt"))
-
-        with self.assertRaises(FileNotFoundError):
-            FPSFilterOperation().run_dataset([self.structure.copy()], params)
-
-    def test_geometry_filter_operation_and_card_roundtrip(self):
-        good = Atoms(
-            symbols=["Si", "Si"],
-            positions=[[0.0, 0.0, 0.0], [2.35, 0.0, 0.0]],
-            cell=np.diag([5.0, 5.0, 5.0]),
-            pbc=True,
-        )
-        bad = good.copy()
-        bad.positions[1] = [0.5, 0.0, 0.0]
-
-        params = GeometryFilterParams(min_pair_distance=1.0, require_finite_cell=True)
-        kept = GeometryFilterOperation().run_dataset([good, bad], params)
-        self.assertEqual(len(kept), 1)
-        self.assertTrue(np.allclose(kept[0].positions, good.positions))
-
-        card = GeometryFilterCard()
-        card.min_pair_frame.set_input_value([1.4])
-        card.min_vpa_frame.set_input_value([5.0])
-        card.max_vpa_frame.set_input_value([80.0])
-        card.require_cell_checkbox.setChecked(True)
-        restored = GeometryFilterCard()
-        restored.from_dict(card.to_dict())
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_shear_matrix_card(self):
-        card = ShearMatrixCard()
-        structure = self.structure.copy()
-        card.organic_checkbox.setChecked(True)
-        card.symmetric_checkbox.setChecked(False)
-        card.xy_frame.set_input_value([1.0, 1.0, 1.0])
-        card.yz_frame.set_input_value([0.0, 0.0, 1.0])
-        card.xz_frame.set_input_value([0.0, 0.0, 1.0])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 1)
-        self.assertFalse(
-            np.allclose(
-                np.array(results[0].get_cell()),
-                np.array(structure.get_cell()),
-                atol=1e-6,
-            )
-        )
-
-    def test_shear_matrix_operation_is_ui_independent(self):
-        params = ShearMatrixParams(
-            xy_range=(1.0, 1.0, 1.0),
-            yz_range=(0.0, 0.0, 1.0),
-            xz_range=(0.0, 0.0, 1.0),
-            symmetric=False,
-        )
-        results = ShearMatrixOperation().run_structure(self.structure.copy(), params)
-
-        self.assertEqual(len(results), 1)
-        self.assertIn("Shr(xy=1%,sym=0)", results[0].info.get("Config_type", ""))
-
-    def test_shear_angle_card(self):
-        card = ShearAngleCard()
-        structure = self.structure.copy()
-        card.organic_checkbox.setChecked(True)
-        card.alpha_frame.set_input_value([1.0, 1.0, 1.0])
-        card.beta_frame.set_input_value([0.0, 0.0, 1.0])
-        card.gamma_frame.set_input_value([0.0, 0.0, 1.0])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 1)
-        original_angles = np.array(structure.cell.cellpar()[3:])
-        new_angles = np.array(results[0].cell.cellpar()[3:])
-        self.assertFalse(np.allclose(new_angles, original_angles, atol=1e-6))
-
-    def test_shear_angle_operation_is_ui_independent(self):
-        params = ShearAngleParams(
-            alpha_range=(1.0, 1.0, 1.0),
-            beta_range=(0.0, 0.0, 1.0),
-            gamma_range=(0.0, 0.0, 1.0),
-        )
-        results = ShearAngleOperation().run_structure(self.structure.copy(), params)
-
-        self.assertEqual(len(results), 1)
-        self.assertIn("Ang(a=1)", results[0].info.get("Config_type", ""))
-
-    def test_random_slab_card(self):
-        card = RandomSlabCard()
-        structure = self.structure.copy()
-        card.h_frame.set_input_value([1, 1, 1])
-        card.k_frame.set_input_value([0, 0, 1])
-        card.l_frame.set_input_value([0, 0, 1])
-        card.layer_frame.set_input_value([1, 1, 1])
-        card.vacuum_frame.set_input_value([0, 0, 1])
-
-        results = card.process_structure(structure)
-        self.assertGreater(len(results), 0)
-        self.assertTrue(all(len(atoms) >= len(structure) for atoms in results))
-
-    def test_random_doping_card(self):
-        card = RandomDopingCard()
-        structure = self.structure.copy()
-        rules = [{
-            "target": "Si",
-            "dopants": {"Ge": 1.0},
-            "use": "count",
-            "count": [1, 1],
-            "concentration": [0.0, 1.0],
-        }]
-        card.rules_widget.from_rules(rules)
-        card.doping_type_combo.setCurrentText("Exact")
-        card.max_atoms_condition_frame.set_input_value([2])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        for atoms in results:
-            self.assertIn("Ge", atoms.get_chemical_symbols())
-
-    def test_random_doping_dopants_accept_bare_element(self):
-        item = DopingRuleItem()
-        item.target_edit.setText("Si")
-        item.dopants_edit.setText("Ge")
-        item.count_botton.setChecked(True)
-        item._on_mode_changed()
-        item.fixed_count_frame.set_input_value([1])
-
-        rule = item.to_rule()
-
-        self.assertEqual(rule["dopants"], {"Ge": 1.0})
-        self.assertEqual(rule["count"], [1, 1])
-        self.assertEqual(rule["count_mode"], "fixed")
-
-    def test_random_doping_count_mode_distinguishes_fixed_and_range(self):
-        structure = Atoms("Si5", positions=np.arange(15, dtype=float).reshape(5, 3), cell=[10, 10, 10], pbc=True)
-
-        fixed = RandomDopingOperation().run_structure(
-            structure,
-            RandomDopingParams(
-                rules=[{"target": "Si", "dopants": {"Ge": 1.0}, "use": "count", "count": [3, 3], "count_mode": "fixed"}],
-                max_structures=3,
-                use_seed=True,
-                seed=1,
-            ),
-        )
-        self.assertTrue(all(atoms.get_chemical_symbols().count("Ge") == 3 for atoms in fixed))
-
-        ranged = RandomDopingOperation().run_structure(
-            structure,
-            RandomDopingParams(
-                rules=[{"target": "Si", "dopants": {"Ge": 1.0}, "use": "count", "count": [1, 3], "count_mode": "random"}],
-                max_structures=10,
-                use_seed=True,
-                seed=1,
-            ),
-        )
-        self.assertTrue(all(1 <= atoms.get_chemical_symbols().count("Ge") <= 3 for atoms in ranged))
-
-    def test_random_doping_operation_is_ui_independent(self):
-        params = RandomDopingParams(
-            rules=[
-                {
-                    "target": "Si",
-                    "dopants": {"Ge": 1.0},
-                    "use": "count",
-                    "count": [1, 1],
-                }
-            ],
-            doping_type="Exact",
-            max_structures=2,
-            use_seed=True,
-            seed=3,
-        )
-        results = RandomDopingOperation().run_structure(self.structure.copy(), params)
-
-        self.assertEqual(len(results), 2)
-        self.assertTrue(all("Ge" in atoms.get_chemical_symbols() for atoms in results))
-
-    def test_parse_composition_accepts_bare_elements(self):
-        self.assertEqual(parse_composition("Ge"), {"Ge": 1.0})
-        self.assertEqual(parse_composition("Ge,C"), {"Ge": 1.0, "C": 1.0})
-        self.assertEqual(parse_composition("Ge:0.7,C"), {"Ge": 0.7, "C": 1.0})
-
-    def test_conditional_replace_operation_is_ui_independent(self):
-        results = ConditionalReplaceOperation().run_structure(
-            self.structure.copy(),
-            ConditionalReplaceParams(
-                target="Si",
-                replacements="Ge:1",
-                condition="all",
-                seed=1,
-                mode=1,
-            ),
-        )
-
-        self.assertEqual(len(results), 1)
-        self.assertTrue(all(symbol == "Ge" for symbol in results[0].get_chemical_symbols()))
-        self.assertIn("Repl(Si->Ge)", results[0].info.get("Config_type", ""))
-
-    def test_conditional_replace_card_roundtrip(self):
-        card = ConditionalReplaceCard()
-        card.target_edit.setText("Si")
-        card.replacements_edit.setText("Ge:0.5,C:0.5")
-        card.condition_edit.setText("z>=0")
-        card.seed_frame.set_input_value([9])
-        card.mode_combo.setCurrentIndex(1)
-
-        restored = ConditionalReplaceCard()
-        restored.from_dict(card.to_dict())
-
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_crystal_prototype_builder_card(self):
-        card = CrystalPrototypeBuilderCard()
-        card.structure_combo.setCurrentText("fcc")
-        card.element_edit.setText("Cu")
-        card.a_frame.set_input_value([3.6, 3.6, 0.1])
-        card.auto_supercell_button.setChecked(True)
-        card.manual_supercell_button.setChecked(False)
-        card.max_atoms_frame.set_input_value([64])
-        card.max_output_frame.set_input_value([10])
-
-        results = card.create_operation().generate(card.get_params())
-        self.assertGreaterEqual(len(results), 1)
-        self.assertTrue(all(atoms.pbc.all() for atoms in results))
-        self.assertTrue(all(len(atoms) <= 64 for atoms in results))
-
-    def test_crystal_prototype_builder_operation_is_ui_independent(self):
-        params = CrystalPrototypeBuilderParams(
-            lattice="bcc",
-            element="Fe",
-            a_range=(2.9, 2.9, 0.1),
-            auto_supercell=False,
-            rep=(1, 1, 1),
-            max_outputs=1,
-        )
-        results = CrystalPrototypeBuilderOperation().generate(params)
-
-        self.assertEqual(len(results), 1)
-        self.assertTrue(results[0].pbc.all())
-        self.assertIn("Proto(bcc", results[0].info.get("Config_type", ""))
-
-    def test_crystal_prototype_builder_card_roundtrip(self):
-        card = CrystalPrototypeBuilderCard()
-        card.structure_combo.setCurrentText("hcp")
-        card.element_edit.setText("Mg")
-        card.a_frame.set_input_value([3.1, 3.3, 0.1])
-        card.covera_frame.set_input_value([1.62])
-        card.auto_supercell_button.setChecked(False)
-        card.manual_supercell_button.setChecked(True)
-        card.max_atoms_frame.set_input_value([128])
-        card.rep_frame.set_input_value([2, 3, 4])
-        card.max_output_frame.set_input_value([5])
-
-        restored = CrystalPrototypeBuilderCard()
-        restored.from_dict(card.to_dict())
-
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_composition_sweep_and_random_occupancy_cards(self):
-        base = self.structure.copy()
-        base.info.setdefault("Config_type", "base")
-
-        sweep = CompositionSweepCard()
-        sweep.elements_edit.setText("Co,Ni")
-        sweep.order_combo.setCurrentText("2")
-        sweep.method_combo.setCurrentText("Grid")
-        sweep.step_frame.set_input_value([0.5])
-        sweep.include_endpoints_checkbox.setChecked(True)
-        sweep.minfrac_frame.set_input_value([0.0])
-        sweep.max_output_frame.set_input_value([10])
-
-        swept = sweep.process_structure(base)
-        self.assertEqual(len(swept), 3)
-        self.assertTrue(
-            all("Comp(" in str(atoms.info.get("Config_type", "")) for atoms in swept)
-        )
-
-        occ = RandomOccupancyCard()
-        occ.source_combo.setCurrentText("Auto (Comp tag)")
-        occ.mode_combo.setCurrentText("Exact")
-        occ.samples_frame.set_input_value([1])
-
-        occupied = []
-        for atoms in swept:
-            occupied.extend(occ.process_structure(atoms))
-        self.assertEqual(len(occupied), len(swept))
-        for atoms in occupied:
-            syms = set(atoms.get_chemical_symbols())
-            self.assertTrue(syms.issubset({"Co", "Ni"}))
-
-    def test_composition_and_occupancy_operations_are_ui_independent(self):
-        base = self.structure.copy()
-        base.info.setdefault("Config_type", "base")
-
-        sweep_params = CompositionSweepParams(
-            elements="Co,Ni",
-            order="2",
-            method="Grid",
-            step=0.5,
-            include_endpoints=True,
-            max_outputs=3,
-        )
-        swept = CompositionSweepOperation().run_structure(base, sweep_params)
-
-        self.assertEqual(len(swept), 3)
-        self.assertTrue(all("Comp(" in atoms.info.get("Config_type", "") for atoms in swept))
-
-        occ_params = RandomOccupancyParams(
-            source="Auto (Comp tag)",
-            mode="Exact",
-            samples=1,
-            use_seed=True,
-            seed=5,
-        )
-        occupied = RandomOccupancyOperation().run_structure(swept[0], occ_params)
-
-        self.assertEqual(len(occupied), 1)
-        self.assertTrue(set(occupied[0].get_chemical_symbols()).issubset({"Co", "Ni"}))
-
-    def test_composition_gradient_operation_and_card_roundtrip(self):
-        base = Atoms(
-            symbols=["Ni"] * 8,
-            positions=[[float(i), 0.0, 0.0] for i in range(8)],
-            cell=np.diag([8.0, 2.0, 2.0]),
-            pbc=[True, False, False],
-        )
-        params = CompositionGradientParams(
-            elements="Ni,Co",
-            start_composition="Ni:1,Co:0",
-            end_composition="Ni:0,Co:1",
-            axis="x",
-            bins=4,
-            samples=2,
-            use_seed=True,
-            seed=3,
-        )
-        results = CompositionGradientOperation().run_structure(base, params)
-        self.assertEqual(len(results), 2)
-        self.assertTrue(all("CompGrad(ax=x,b=4" in str(atoms.info.get("Config_type", "")) for atoms in results))
-        self.assertEqual(results[0].get_chemical_symbols()[:2], ["Ni", "Ni"])
-        self.assertEqual(results[0].get_chemical_symbols()[-2:], ["Co", "Co"])
-
-        card = CompositionGradientCard()
-        card.elements_edit.setText("Ni,Co")
-        card.bins_frame.set_input_value([4])
-        card.samples_frame.set_input_value([2])
-        card.seed_checkbox.setChecked(True)
-        card.seed_frame.set_input_value([3])
-        restored = CompositionGradientCard()
-        restored.from_dict(card.to_dict())
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_composition_sweep_quaternary_quinary(self):
-        base = self.structure.copy()
-        base.info.setdefault("Config_type", "base")
-
-        sweep4 = CompositionSweepCard()
-        sweep4.elements_edit.setText("Co,Cr,Ni,Al,Fe")
-        sweep4.order_combo.setCurrentText("4")
-        sweep4.method_combo.setCurrentText("Sobol")
-        sweep4.n_points_frame.set_input_value([8])
-        sweep4.max_output_frame.set_input_value([8])
-
-        swept4 = sweep4.process_structure(base)
-        self.assertEqual(len(swept4), 8)
-        for atoms in swept4:
-            cfg = str(atoms.info.get("Config_type", ""))
-            comp_tokens = [t.strip() for t in cfg.split("|") if t.strip().startswith("Comp(") and t.strip().endswith(")")]
-            self.assertTrue(comp_tokens)
-            comp_items = [p for p in comp_tokens[-1][5:-1].split(",") if p.strip()]
-            self.assertEqual(len(comp_items), 4)
-
-        sweep5 = CompositionSweepCard()
-        sweep5.elements_edit.setText("Co,Cr,Ni,Al,Fe")
-        sweep5.order_combo.setCurrentText("5")
-        sweep5.method_combo.setCurrentText("Sobol")
-        sweep5.n_points_frame.set_input_value([6])
-        sweep5.max_output_frame.set_input_value([6])
-
-        swept5 = sweep5.process_structure(base)
-        self.assertEqual(len(swept5), 6)
-        for atoms in swept5:
-            cfg = str(atoms.info.get("Config_type", ""))
-            comp_tokens = [t.strip() for t in cfg.split("|") if t.strip().startswith("Comp(") and t.strip().endswith(")")]
-            self.assertTrue(comp_tokens)
-            comp_items = [p for p in comp_tokens[-1][5:-1].split(",") if p.strip()]
-            self.assertEqual(len(comp_items), 5)
-
-    def test_magnetic_order_card_fm_afm(self):
-        proto = CrystalPrototypeBuilderCard()
-        proto.structure_combo.setCurrentText("bcc")
-        proto.element_edit.setText("Fe")
-        proto.a_frame.set_input_value([2.9, 2.9, 0.1])
-        proto.manual_supercell_button.setChecked(True)
-        proto.auto_supercell_button.setChecked(False)
-        proto.rep_frame.set_input_value([2, 2, 2])
-        proto.max_output_frame.set_input_value([1])
-        base = proto.create_operation().generate(proto.get_params())[0]
-
-        card = MagneticOrderCard()
-        card.map_edit.setText("Fe:2.2")
-        card.fm_checkbox.setChecked(True)
-        card.afm_checkbox.setChecked(True)
-        card.kvec_combo.setCurrentText("100")
-        card.pm_checkbox.setChecked(False)
-
-        results = card.process_structure(base)
-        self.assertEqual(len(results), 2)
-        fm = [a for a in results if "MagFM" in str(a.info.get("Config_type", ""))][0]
-        afm = [a for a in results if "MagAFM100" in str(a.info.get("Config_type", ""))][0]
-
-        fm_m = np.array(fm.get_initial_magnetic_moments(), dtype=float)
-        afm_m = np.array(afm.get_initial_magnetic_moments(), dtype=float)
-        self.assertTrue(np.all(fm_m >= 0))
-        self.assertTrue(np.any(afm_m > 0) and np.any(afm_m < 0))
-
-    def test_magnetic_operations_are_ui_independent(self):
-        structure = self._spin_chain()
-
-        set_result = SetMagneticMomentsOperation().run_structure(
-            structure,
-            SetMagneticMomentsParams(
-                source="Map/default magnitude",
-                format="Non-collinear (vector)",
-                magmom_map="Fe:2.0",
-                axis=[0.0, 0.0, 1.0],
-            ),
-        )[0]
-        self.assertIn("MagSet(map,vec)", str(set_result.info.get("Config_type", "")))
-
-        order_result = MagneticOrderOperation().run_structure(
-            structure,
-            MagneticOrderParams(magmom_map="Fe:2.0", gen_fm=True, gen_afm=False),
-        )[0]
-        self.assertIn("MagFM", str(order_result.info.get("Config_type", "")))
-
-        spiral_result = SpinSpiralOperation().run_structure(
-            set_result,
-            SpinSpiralParams(
-                period_range=[4.0, 4.0, 1.0],
-                phase_range=[0.0, 0.0, 15.0],
-                chirality="Clockwise",
-                max_outputs=1,
-            ),
-        )[0]
-        self.assertIn("Helix(", str(spiral_result.info.get("Config_type", "")))
-
-        folded = FoldedHelixOperation().run_structure(
-            set_result,
-            FoldedHelixParams(
-                half_period_mode="Manual",
-                half_period_layers=[2, 2, 1],
-                angle_step_range=[30.0, 30.0, 15.0],
-                phase_range=[0.0, 0.0, 15.0],
-                max_outputs=1,
-            ),
-        )[0]
-        self.assertIn("FoldedHelix(", str(folded.info.get("Config_type", "")))
-
-        tilt_result = SmallAngleSpinTiltOperation().run_structure(
-            set_result,
-            SmallAngleSpinTiltParams(
-                target_mode="Explicit indices (1-based)",
-                target_indices="2",
-                angle_list="5",
-                include_reference=False,
-            ),
-        )[0]
-        self.assertIn("SpinTilt(i=2,a=5", str(tilt_result.info.get("Config_type", "")))
-
-        rotated = MagneticMomentRotationOperation().run_structure(
-            set_result,
-            MagneticMomentRotationParams(
-                elements="Fe",
-                max_angle=10.0,
-                num_structures=1,
-                use_seed=True,
-                seed=7,
-            ),
-        )[0]
-        moments = np.array(rotated.get_initial_magnetic_moments(), dtype=float)
-        self.assertEqual(moments.shape, (4, 3))
-
-    def test_group_label_card_and_group_afm(self):
-        proto = CrystalPrototypeBuilderCard()
-        proto.structure_combo.setCurrentText("bcc")
-        proto.element_edit.setText("Fe")
-        proto.a_frame.set_input_value([2.9, 2.9, 0.1])
-        proto.manual_supercell_button.setChecked(True)
-        proto.auto_supercell_button.setChecked(False)
-        proto.rep_frame.set_input_value([2, 2, 2])
-        proto.max_output_frame.set_input_value([1])
-        base = proto.create_operation().generate(proto.get_params())[0]
-
-        gl = GroupLabelCard()
-        gl.mode_combo.setCurrentText("k-vector layers (recommended)")
-        gl.kvec_combo.setCurrentText("111")
-        gl.group_a_edit.setText("A")
-        gl.group_b_edit.setText("B")
-        labeled = gl.process_structure(base)[0]
-        self.assertIn("group", labeled.arrays)
-        groups = set(str(g) for g in labeled.arrays["group"])
-        self.assertTrue({"A", "B"}.issubset(groups))
-
-        mag = MagneticOrderCard()
-        mag.format_combo.setCurrentText("Collinear (scalar)")
-        mag.map_edit.setText("Fe:2.2")
-        mag.fm_checkbox.setChecked(False)
-        mag.afm_checkbox.setChecked(True)
-        mag.afm_mode_combo.setCurrentText("group A/B")
-        mag.group_a_edit.setText("A")
-        mag.group_b_edit.setText("B")
-        mag.pm_checkbox.setChecked(False)
-        res = mag.process_structure(labeled)
-        self.assertEqual(len(res), 1)
-        afm = res[0]
-        m = np.array(afm.get_initial_magnetic_moments(), dtype=float)
-        self.assertTrue(np.any(m > 0) and np.any(m < 0))
-
-    def test_group_label_operation_is_ui_independent(self):
-        base = CrystalPrototypeBuilderOperation().generate(
-            CrystalPrototypeBuilderParams(
-                lattice="bcc",
-                element="Fe",
-                a_range=(2.9, 2.9, 0.1),
-                auto_supercell=False,
-                rep=(2, 2, 2),
-                max_outputs=1,
-            )
-        )[0]
-
-        labeled = GroupLabelOperation().run_structure(
-            base,
-            GroupLabelParams(
-                mode="k-vector layers (recommended)",
-                kvec="111",
-                group_a="up",
-                group_b="down",
-            ),
-        )[0]
-
-        self.assertIn("group", labeled.arrays)
-        self.assertEqual(set(str(value) for value in labeled.arrays["group"]), {"up", "down"})
-        self.assertIn("Grp(k111,up/down)", labeled.info.get("Config_type", ""))
-
-    def test_group_label_card_roundtrip(self):
-        card = GroupLabelCard()
-        card.mode_combo.setCurrentText("fractional parity")
-        card.kvec_combo.setCurrentText("110")
-        card.group_a_edit.setText("alpha")
-        card.group_b_edit.setText("beta")
-        card.overwrite_checkbox.setChecked(False)
-
-        restored = GroupLabelCard()
-        restored.from_dict(card.to_dict())
-
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_magnetic_order_card_noncollinear_pm(self):
-        proto = CrystalPrototypeBuilderCard()
-        proto.structure_combo.setCurrentText("fcc")
-        proto.element_edit.setText("Ni")
-        proto.a_frame.set_input_value([3.5, 3.5, 0.1])
-        proto.manual_supercell_button.setChecked(True)
-        proto.auto_supercell_button.setChecked(False)
-        proto.rep_frame.set_input_value([2, 2, 2])
-        proto.max_output_frame.set_input_value([1])
-        base = proto.create_operation().generate(proto.get_params())[0]
-
-        card = MagneticOrderCard()
-        card.format_combo.setCurrentText("Non-collinear (vector)")
-        card.axis_frame.set_input_value([0.0, 0.0, 1.0])
-        card.map_edit.setText("Ni:0.6")
-        card.fm_checkbox.setChecked(False)
-        card.afm_checkbox.setChecked(False)
-        card.pm_checkbox.setChecked(True)
-        card.pm_count_frame.set_input_value([2])
-        card.pm_direction_combo.setCurrentText("sphere")
-        card.pm_balanced_checkbox.setChecked(True)
-        card.seed_checkbox.setChecked(True)
-        card.seed_frame.set_input_value([123])
-
-        results = card.process_structure(base)
-        self.assertEqual(len(results), 2)
-        for atoms in results:
-            m = np.array(atoms.get_initial_magnetic_moments(), dtype=float)
-            self.assertEqual(m.ndim, 2)
-            self.assertEqual(m.shape[1], 3)
-            self.assertTrue(np.any(np.linalg.norm(m, axis=1) > 0))
-            self.assertIn("MagPMnc", str(atoms.info.get("Config_type", "")))
-
-    def test_spin_spiral_card_periods_and_chirality(self):
-        structure = self._spin_chain()
-
-        card = SpinSpiralCard()
-        card.source_combo.setCurrentText("Map/default magnitude")
-        card.map_edit.setText("Fe:2.0")
-        card.period_frame.set_input_value([2.0, 4.0, 2.0])
-        card.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        card.mz_frame.set_input_value([0.0, 0.0, 0.1])
-        card.chirality_combo.setCurrentText("Both")
-        card.max_output_frame.set_input_value([10])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 4)
-
-        for atoms in results:
-            moments = np.array(atoms.get_initial_magnetic_moments(), dtype=float)
-            self.assertEqual(moments.shape, (4, 3))
-            self.assertTrue(np.allclose(np.linalg.norm(moments, axis=1), 2.0, atol=1e-6))
-            self.assertTrue(np.allclose(moments[:, 2], 0.0, atol=1e-6))
-            self.assertIn("Helix(", str(atoms.info.get("Config_type", "")))
-
-        cw = next(
-            atoms for atoms in results
-            if "L=4" in str(atoms.info.get("Config_type", "")) and "chi=cw" in str(atoms.info.get("Config_type", ""))
-        )
-        ccw = next(
-            atoms for atoms in results
-            if "L=4" in str(atoms.info.get("Config_type", "")) and "chi=ccw" in str(atoms.info.get("Config_type", ""))
-        )
-        cw_m = np.array(cw.get_initial_magnetic_moments(), dtype=float)
-        ccw_m = np.array(ccw.get_initial_magnetic_moments(), dtype=float)
-        self.assertAlmostEqual(cw_m[1, 0], -ccw_m[1, 0], places=6)
-        self.assertAlmostEqual(cw_m[1, 1], ccw_m[1, 1], places=6)
-
-    def test_spin_spiral_card_mz_scan_generates_multiple_conical_states(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        card = SpinSpiralCard()
-        card.period_frame.set_input_value([6.0, 6.0, 1.0])
-        card.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        card.mz_frame.set_input_value([0.0, 0.8, 0.4])
-        card.chirality_combo.setCurrentText("Clockwise")
-        card.max_output_frame.set_input_value([10])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 3)
-
-        expected_mz_values = [0.0, 0.4, 0.8]
-        for atoms, expected_mz in zip(results, expected_mz_values):
-            moments = np.array(atoms.get_initial_magnetic_moments(), dtype=float)
-            norms = np.linalg.norm(moments, axis=1)
-            self.assertTrue(np.allclose(norms, 2.0, atol=1e-6))
-            self.assertTrue(np.allclose(moments[:, 2], norms * expected_mz, atol=1e-6))
-            tag = str(atoms.info.get("Config_type", ""))
-            self.assertIn(f"mz={expected_mz:.6g}", tag)
-            if expected_mz == 0.0:
-                self.assertIn("Helix(", tag)
-            else:
-                self.assertIn("Spiral(", tag)
-
-    def test_spin_spiral_card_commensurate_period_filter(self):
-        structure = self._spin_chain()
-
-        card = SpinSpiralCard()
-        card.source_combo.setCurrentText("Map/default magnitude")
-        card.map_edit.setText("Fe:2.0")
-        card.period_frame.set_input_value([4.0, 8.0, 2.0])
-        card.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        card.mz_frame.set_input_value([0.0, 0.0, 0.1])
-        card.chirality_combo.setCurrentText("Both")
-        card.commensurate_checkbox.setChecked(True)
-        card.max_output_frame.set_input_value([10])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 4)
-
-        tags = [str(atoms.info.get("Config_type", "")) for atoms in results]
-        self.assertTrue(all(("L=4" in tag or "L=8" in tag) for tag in tags))
-        self.assertFalse(any("L=6" in tag for tag in tags))
-
-        data = card.to_dict()
-        restored = SpinSpiralCard()
-        restored.from_dict(data)
-        self.assertTrue(restored.commensurate_checkbox.isChecked())
-
-    def test_spin_spiral_card_commensurate_period_suggests_supercell(self):
-        structure = self._spin_chain()
-        suggestion = SpinSpiralCard._suggest_supercell_multipliers(
-            [6.0, 10.0],
-            structure=structure,
-            axis=np.array([0.0, 0.0, 1.0]),
-        )
-        self.assertIsNotNone(suggestion)
-        suggested_period, multipliers = suggestion
-        self.assertAlmostEqual(suggested_period, 6.0)
-        self.assertEqual(multipliers, [1, 1, 3])
-
-    def test_spin_spiral_card_commensurate_period_discovery_uses_continuous_range(self):
-        structure = self._spin_chain()
-        structure.set_cell(np.diag([6.0, 6.0, 31.0]))
-
-        card = SpinSpiralCard()
-        card.source_combo.setCurrentText("Map/default magnitude")
-        card.map_edit.setText("Fe:2.0")
-        card.period_frame.set_input_value([15.0, 16.0, 1.0])
-        card.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        card.mz_frame.set_input_value([0.0, 0.0, 0.1])
-        card.chirality_combo.setCurrentText("Both")
-        card.commensurate_checkbox.setChecked(True)
-        card.max_output_frame.set_input_value([10])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        tags = [str(atoms.info.get("Config_type", "")) for atoms in results]
-        self.assertTrue(all("L=15.5" in tag for tag in tags))
-
-    def test_spin_spiral_card_layer_locked_phase_keeps_each_layer_rigid(self):
-        structure = self._spin_bilayer_chain()
-        structure.set_initial_magnetic_moments([2.0] * len(structure))
-
-        continuous = SpinSpiralCard()
-        continuous.period_frame.set_input_value([4.0, 4.0, 1.0])
-        continuous.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        continuous.mz_frame.set_input_value([0.0, 0.0, 0.1])
-        continuous.chirality_combo.setCurrentText("Clockwise")
-
-        layer_locked = SpinSpiralCard()
-        layer_locked.period_frame.set_input_value([4.0, 4.0, 1.0])
-        layer_locked.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        layer_locked.mz_frame.set_input_value([0.0, 0.0, 0.1])
-        layer_locked.chirality_combo.setCurrentText("Clockwise")
-        layer_locked.phase_mode_combo.setCurrentText("Layer-locked")
-        layer_locked.layer_tol_frame.set_input_value([0.05])
-
-        continuous_result = continuous.process_structure(structure)[0]
-        layer_locked_result = layer_locked.process_structure(structure)[0]
-
-        continuous_moments = np.array(continuous_result.get_initial_magnetic_moments(), dtype=float)
-        layer_locked_moments = np.array(layer_locked_result.get_initial_magnetic_moments(), dtype=float)
-
-        self.assertFalse(np.allclose(continuous_moments[0], continuous_moments[1], atol=1e-6))
-        self.assertTrue(np.allclose(layer_locked_moments[0], layer_locked_moments[1], atol=1e-6))
-        self.assertTrue(np.allclose(layer_locked_moments[2], layer_locked_moments[3], atol=1e-6))
-        self.assertTrue(np.allclose(layer_locked_moments[4], layer_locked_moments[5], atol=1e-6))
-        self.assertTrue(np.allclose(layer_locked_moments[6], layer_locked_moments[7], atol=1e-6))
-        self.assertIn("pm=layer", str(layer_locked_result.info.get("Config_type", "")))
-
-    def test_spin_spiral_card_existing_magnitudes_and_roundtrip(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([1.0, 2.0, 3.0, 4.0])
-
-        card = SpinSpiralCard()
-        card.axis_frame.set_input_value([0.0, 0.0, 1.0])
-        card.period_frame.set_input_value([6.0, 6.0, 1.0])
-        card.phase_frame.set_input_value([30.0, 30.0, 15.0])
-        card.mz_frame.set_input_value([0.5, 0.5, 0.1])
-        card.chirality_combo.setCurrentText("Clockwise")
-        card.source_combo.setCurrentText("Existing initial magmoms")
-        card.apply_edit.setText("Fe")
-        card.max_output_frame.set_input_value([5])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 1)
-        moments = np.array(results[0].get_initial_magnetic_moments(), dtype=float)
-        norms = np.linalg.norm(moments, axis=1)
-        self.assertTrue(np.allclose(norms, [1.0, 2.0, 3.0, 4.0], atol=1e-6))
-        self.assertTrue(np.allclose(moments[:, 2], norms * 0.5, atol=1e-6))
-        self.assertIn("Spiral(", str(results[0].info.get("Config_type", "")))
-        self.assertIn("chi=cw", str(results[0].info.get("Config_type", "")))
-
-        data = card.to_dict()
-        restored = SpinSpiralCard()
-        restored.from_dict(data)
-        self.assertEqual(restored.parameter_mode_combo.currentText(), "Period (L_D)")
-        self.assertEqual(restored.chirality_combo.currentText(), "Clockwise")
-        self.assertEqual(restored.source_combo.currentText(), "Existing initial magmoms")
-        self.assertEqual(restored.apply_edit.text(), "Fe")
-        self.assertEqual(restored.period_frame.get_input_value(), [6.0, 6.0, 1.0])
-        self.assertEqual(restored.mz_frame.get_input_value(), [0.5, 0.5, 0.1])
-        self.assertEqual(restored.phase_mode_combo.currentText(), "Continuous by position")
-        self.assertEqual(restored.layer_tol_frame.get_input_value(), [0.05])
-        self.assertFalse(restored.map_label.isVisible())
-        self.assertFalse(restored.default_label.isVisible())
-
-        legacy = SpinSpiralCard()
-        legacy.from_dict({"check_state": True, "mz": [0.25]})
-        self.assertEqual(legacy.mz_frame.get_input_value(), [0.25, 0.25, 0.1])
-
-    def test_spin_spiral_card_layer_locked_roundtrip(self):
-        card = SpinSpiralCard()
-        card.phase_mode_combo.setCurrentText("Layer-locked")
-        card.layer_tol_frame.set_input_value([0.08])
-
-        data = card.to_dict()
-        restored = SpinSpiralCard()
-        restored.from_dict(data)
-
-        self.assertEqual(restored.phase_mode_combo.currentText(), "Layer-locked")
-        self.assertEqual(restored.layer_tol_frame.get_input_value(), [0.08])
-        self.assertFalse(restored.layer_tol_label.isHidden())
-        self.assertTrue(restored.layer_tol_frame.isEnabled())
-
-    def test_spin_spiral_card_angle_gradient_mode_matches_period_mode(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        by_period = SpinSpiralCard()
-        by_period.period_frame.set_input_value([4.0, 4.0, 1.0])
-        by_period.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        by_period.chirality_combo.setCurrentText("Counterclockwise")
-        period_result = by_period.process_structure(structure)[0]
-
-        by_gradient = SpinSpiralCard()
-        by_gradient.parameter_mode_combo.setCurrentText("Angle gradient (deg/A)")
-        by_gradient.angle_gradient_frame.set_input_value([90.0, 90.0, 1.0])
-        by_gradient.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        by_gradient.chirality_combo.setCurrentText("Counterclockwise")
-        gradient_result = by_gradient.process_structure(structure)[0]
-
-        self.assertTrue(
-            np.allclose(
-                np.array(period_result.get_initial_magnetic_moments(), dtype=float),
-                np.array(gradient_result.get_initial_magnetic_moments(), dtype=float),
-                atol=1e-6,
-            )
-        )
-
-        data = by_gradient.to_dict()
-        restored = SpinSpiralCard()
-        restored.from_dict(data)
-        self.assertEqual(restored.parameter_mode_combo.currentText(), "Angle gradient (deg/A)")
-        self.assertEqual(restored.angle_gradient_frame.get_input_value(), [90.0, 90.0, 1.0])
-
-    def test_folded_helix_card_layer_pattern(self):
-        structure = self._spin_bilayer_chain()
-
-        card = FoldedHelixCard()
-        card.source_combo.setCurrentText("Map/default magnitude")
-        card.map_edit.setText("Fe:2.0")
-        card.layer_axis_frame.set_input_value([0.0, 0.0, 1.0])
-        card.plane_normal_frame.set_input_value([0.0, 0.0, 1.0])
-        card.layer_tol_frame.set_input_value([0.05])
-        card.half_period_mode_combo.setCurrentText("Manual")
-        card.half_period_frame.set_input_value([2, 2, 1])
-        card.angle_step_frame.set_input_value([30.0, 30.0, 15.0])
-        card.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        card.sequence_combo.setCurrentText("Clockwise then counterclockwise")
-        card.max_output_frame.set_input_value([4])
-
-        result = card.process_structure(structure)[0]
-        moments = np.array(result.get_initial_magnetic_moments(), dtype=float)
-        self.assertEqual(moments.shape, (8, 3))
-        self.assertTrue(np.allclose(np.linalg.norm(moments, axis=1), 2.0, atol=1e-6))
-        self.assertTrue(np.allclose(moments[:, 2], 0.0, atol=1e-6))
-        self.assertTrue(np.allclose(moments[0], moments[1], atol=1e-6))
-        self.assertTrue(np.allclose(moments[2], moments[3], atol=1e-6))
-        self.assertTrue(np.allclose(moments[4], moments[5], atol=1e-6))
-        self.assertTrue(np.allclose(moments[6], moments[7], atol=1e-6))
-
-        e1, e2, _ = orthonormal_frame(np.array([0.0, 0.0, 1.0], dtype=float))
-        coeff_1 = moments @ e1 / 2.0
-        coeff_2 = moments @ e2 / 2.0
-        phase_deg = np.rad2deg(np.arctan2(coeff_2, coeff_1))
-        expected = np.array([0.0, 0.0, -30.0, -30.0, -60.0, -60.0, -30.0, -30.0], dtype=float)
-        wrapped = ((phase_deg - expected + 180.0) % 360.0) - 180.0
-        self.assertTrue(np.allclose(wrapped, 0.0, atol=1e-6))
-        self.assertIn("FoldedHelix(", str(result.info.get("Config_type", "")))
-
-    def test_folded_helix_card_auto_half_period_from_layer_count(self):
-        structure = self._spin_bilayer_chain()
-        structure.set_initial_magnetic_moments([2.0] * len(structure))
-
-        card = FoldedHelixCard()
-        card.layer_axis_frame.set_input_value([0.0, 0.0, 1.0])
-        card.plane_normal_frame.set_input_value([0.0, 0.0, 1.0])
-        card.layer_tol_frame.set_input_value([0.05])
-        card.half_period_mode_combo.setCurrentText("Auto from layer count")
-        card.angle_step_frame.set_input_value([30.0, 30.0, 15.0])
-        card.phase_frame.set_input_value([0.0, 0.0, 15.0])
-        card.sequence_combo.setCurrentText("Clockwise then counterclockwise")
-
-        result = card.process_structure(structure)[0]
-        moments = np.array(result.get_initial_magnetic_moments(), dtype=float)
-        e1, e2, _ = orthonormal_frame(np.array([0.0, 0.0, 1.0], dtype=float))
-        coeff_1 = moments @ e1 / 2.0
-        coeff_2 = moments @ e2 / 2.0
-        phase_deg = np.rad2deg(np.arctan2(coeff_2, coeff_1))
-        expected = np.array([0.0, 0.0, -30.0, -30.0, -30.0, -30.0, 0.0, 0.0], dtype=float)
-        wrapped = ((phase_deg - expected + 180.0) % 360.0) - 180.0
-        self.assertTrue(np.allclose(wrapped, 0.0, atol=1e-6))
-        self.assertTrue(np.allclose(moments[0], moments[-1], atol=1e-6))
-        self.assertFalse(card.half_period_frame.isEnabled())
-
-    def test_folded_helix_card_both_sequences_and_roundtrip(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([1.0, 2.0, 3.0, 4.0])
-
-        card = FoldedHelixCard()
-        card.layer_axis_frame.set_input_value([0.0, 0.0, 1.0])
-        card.plane_normal_frame.set_input_value([0.0, 0.0, 1.0])
-        card.half_period_mode_combo.setCurrentText("Manual")
-        card.half_period_frame.set_input_value([2, 2, 1])
-        card.angle_step_frame.set_input_value([45.0, 45.0, 15.0])
-        card.phase_frame.set_input_value([15.0, 15.0, 15.0])
-        card.sequence_combo.setCurrentText("Both")
-        card.apply_edit.setText("Fe")
-        card.max_output_frame.set_input_value([10])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        tags = [str(atoms.info.get("Config_type", "")) for atoms in results]
-        self.assertTrue(any("seq=cw-ccw" in tag for tag in tags))
-        self.assertTrue(any("seq=ccw-cw" in tag for tag in tags))
-        for atoms in results:
-            moments = np.array(atoms.get_initial_magnetic_moments(), dtype=float)
-            self.assertEqual(moments.shape, (4, 3))
-            self.assertTrue(np.allclose(np.linalg.norm(moments, axis=1), [1.0, 2.0, 3.0, 4.0], atol=1e-6))
-
-        data = card.to_dict()
-        restored = FoldedHelixCard()
-        restored.from_dict(data)
-        self.assertEqual(restored.half_period_mode_combo.currentText(), "Manual")
-        self.assertEqual(restored.sequence_combo.currentText(), "Both")
-        self.assertEqual(restored.half_period_frame.get_input_value(), [2, 2, 1])
-        self.assertEqual(restored.angle_step_frame.get_input_value(), [45.0, 45.0, 15.0])
-        self.assertEqual(restored.phase_frame.get_input_value(), [15.0, 15.0, 15.0])
-        self.assertEqual(restored.apply_edit.text(), "Fe")
-        self.assertFalse(restored.map_label.isVisible())
-
-    def test_small_angle_spin_tilt_card_reference_and_explicit_index(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        card = SmallAngleSpinTiltCard()
-        card.target_mode_combo.setCurrentText("Explicit indices (1-based)")
-        card.target_indices_edit.setText("2")
-        card.angle_edit.setText("1,5")
-        card.include_reference_checkbox.setChecked(True)
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 3)
-
-        reference = np.array(results[0].get_initial_magnetic_moments(), dtype=float)
-        tilted_1 = np.array(results[1].get_initial_magnetic_moments(), dtype=float)
-        tilted_5 = np.array(results[2].get_initial_magnetic_moments(), dtype=float)
-
-        self.assertTrue(np.allclose(reference, np.tile([0.0, 0.0, 2.0], (4, 1)), atol=1e-6))
-        self.assertTrue(np.allclose(tilted_1[[0, 2, 3]], reference[[0, 2, 3]], atol=1e-6))
-        self.assertTrue(np.allclose(tilted_5[[0, 2, 3]], reference[[0, 2, 3]], atol=1e-6))
-        self.assertAlmostEqual(np.linalg.norm(tilted_1[1]), 2.0, places=6)
-        self.assertAlmostEqual(np.linalg.norm(tilted_5[1]), 2.0, places=6)
-        self.assertAlmostEqual(tilted_1[1, 0], 2.0 * np.sin(np.deg2rad(1.0)), places=6)
-        self.assertAlmostEqual(tilted_1[1, 2], 2.0 * np.cos(np.deg2rad(1.0)), places=6)
-        self.assertAlmostEqual(tilted_5[1, 0], 2.0 * np.sin(np.deg2rad(5.0)), places=6)
-        self.assertAlmostEqual(tilted_5[1, 2], 2.0 * np.cos(np.deg2rad(5.0)), places=6)
-        self.assertIn("SpinTiltRef", str(results[0].info.get("Config_type", "")))
-        self.assertIn("SpinTilt(i=2,a=5", str(results[2].info.get("Config_type", "")))
-        self.assertFalse(card.map_label.isVisible())
-        self.assertFalse(card.default_label.isVisible())
-
-    def test_small_angle_spin_tilt_global_tilt(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        results = SmallAngleSpinTiltOperation().run_structure(
-            structure,
-            SmallAngleSpinTiltParams(
-                canting_mode="Global tilt",
-                angle_list="5",
-                include_reference=False,
-            ),
-        )
-        self.assertEqual(len(results), 1)
-        moments = np.array(results[0].get_initial_magnetic_moments(), dtype=float)
-        self.assertTrue(np.allclose(moments[:, 0], 2.0 * np.sin(np.deg2rad(5.0)), atol=1e-6))
-        self.assertTrue(np.allclose(moments[:, 2], 2.0 * np.cos(np.deg2rad(5.0)), atol=1e-6))
-        self.assertIn("SpinTiltG(a=5,sg=pos)", str(results[0].info.get("Config_type", "")))
-
-    def test_small_angle_spin_tilt_card_map_source_roundtrip_and_limit(self):
-        structure = self._spin_chain()
-
-        card = SmallAngleSpinTiltCard()
-        card.source_combo.setCurrentText("Map/default magnitude")
-        card.map_edit.setText("Fe:2.5")
-        card.target_mode_combo.setCurrentText("All eligible atoms")
-        card.angle_edit.setText("1,2")
-        card.sign_combo.setCurrentText("Both (+/- pair)")
-        card.include_reference_checkbox.setChecked(False)
-        card.max_output_frame.set_input_value([3])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 3)
-        for atoms in results:
-            moments = np.array(atoms.get_initial_magnetic_moments(), dtype=float)
-            self.assertEqual(moments.shape, (4, 3))
-            self.assertTrue(np.allclose(np.linalg.norm(moments, axis=1), 2.5, atol=1e-6))
-            self.assertIn("SpinTilt(", str(atoms.info.get("Config_type", "")))
-            self.assertRegex(str(atoms.info.get("Config_type", "")), r"sg=(pos|neg)")
-
-        data = card.to_dict()
-        restored = SmallAngleSpinTiltCard()
-        restored.from_dict(data)
-        self.assertEqual(restored.source_combo.currentText(), "Map/default magnitude")
-        self.assertEqual(restored.target_mode_combo.currentText(), "All eligible atoms")
-        self.assertEqual(restored.angle_edit.text(), "1,2")
-        self.assertEqual(restored.sign_combo.currentText(), "Both (+/- pair)")
-        self.assertFalse(restored.include_reference_checkbox.isChecked())
-        self.assertEqual(restored.max_output_frame.get_input_value(), [3])
-
-    def test_small_angle_spin_tilt_card_atom_pair_canting(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        card = SmallAngleSpinTiltCard()
-        card.canting_mode_combo.setCurrentText("Atom pair canting")
-        card.pair_left_edit.setText("1")
-        card.pair_right_edit.setText("2")
-        card.angle_edit.setText("10")
-        card.sign_combo.setCurrentText("Both (+/- pair)")
-        card.include_reference_checkbox.setChecked(False)
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-
-        pos = next(a for a in results if "sg=pos" in str(a.info.get("Config_type", "")))
-        neg = next(a for a in results if "sg=neg" in str(a.info.get("Config_type", "")))
-        pos_m = np.array(pos.get_initial_magnetic_moments(), dtype=float)
-        neg_m = np.array(neg.get_initial_magnetic_moments(), dtype=float)
-
-        self.assertAlmostEqual(pos_m[0, 0], 2.0 * np.sin(np.deg2rad(5.0)), places=6)
-        self.assertAlmostEqual(pos_m[1, 0], -2.0 * np.sin(np.deg2rad(5.0)), places=6)
-        self.assertAlmostEqual(neg_m[0, 0], -2.0 * np.sin(np.deg2rad(5.0)), places=6)
-        self.assertAlmostEqual(neg_m[1, 0], 2.0 * np.sin(np.deg2rad(5.0)), places=6)
-        self.assertIn("SpinPair(i=1,j=2,a=10,sg=pos)", str(pos.info.get("Config_type", "")))
-
-    def test_small_angle_spin_tilt_card_group_pair_canting(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-        structure.new_array("group", np.array(["A", "A", "B", "B"], dtype="<U1"))
-
-        card = SmallAngleSpinTiltCard()
-        card.canting_mode_combo.setCurrentText("Group pair canting")
-        card.group_a_edit.setText("A")
-        card.group_b_edit.setText("B")
-        card.angle_edit.setText("6")
-        card.sign_combo.setCurrentText("Positive only")
-        card.include_reference_checkbox.setChecked(False)
-
-        result = card.process_structure(structure)[0]
-        moments = np.array(result.get_initial_magnetic_moments(), dtype=float)
-        expected = 2.0 * np.sin(np.deg2rad(3.0))
-        self.assertTrue(np.allclose(moments[[0, 1], 0], expected, atol=1e-6))
-        self.assertTrue(np.allclose(moments[[2, 3], 0], -expected, atol=1e-6))
-        self.assertIn("SpinPairG(A=A,B=B,a=6,sg=pos)", str(result.info.get("Config_type", "")))
-
-    def test_small_angle_spin_tilt_card_auto_neighbor_shell_pair(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        card = SmallAngleSpinTiltCard()
-        card.canting_mode_combo.setCurrentText("Atom pair canting")
-        card.pair_source_combo.setCurrentText("Auto by neighbor shell")
-        card.pair_shell_frame.set_input_value([1])
-        card.pair_tol_frame.set_input_value([0.02])
-        card.angle_edit.setText("4")
-        card.sign_combo.setCurrentText("Positive only")
-        card.include_reference_checkbox.setChecked(False)
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 3)
-        tags = [str(a.info.get("Config_type", "")) for a in results]
-        self.assertTrue(any("SpinPair(i=1,j=2,a=4,sg=pos)" in tag for tag in tags))
-        self.assertTrue(any("SpinPair(i=2,j=3,a=4,sg=pos)" in tag for tag in tags))
-        self.assertTrue(any("SpinPair(i=3,j=4,a=4,sg=pos)" in tag for tag in tags))
-
-        first = next(a for a in results if "SpinPair(i=1,j=2,a=4,sg=pos)" in str(a.info.get("Config_type", "")))
-        moments = np.array(first.get_initial_magnetic_moments(), dtype=float)
-        expected = 2.0 * np.sin(np.deg2rad(2.0))
-        self.assertAlmostEqual(moments[0, 0], expected, places=6)
-        self.assertAlmostEqual(moments[1, 0], -expected, places=6)
-
-    def test_small_angle_spin_tilt_card_auto_neighbor_shell_filters_and_roundtrip(self):
-        structure = Atoms(
-            symbols=["Fe", "Co", "Fe", "Co"],
-            positions=[
-                [0.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0],
-                [0.0, 0.0, 2.0],
-                [0.0, 0.0, 3.0],
-            ],
-            cell=np.diag([6.0, 6.0, 8.0]),
-            pbc=[False, False, True],
-        )
-        structure.info["Config_type"] = "FeCo_chain"
-        structure.new_array("group", np.array(["A", "A", "B", "B"], dtype="<U1"))
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        card = SmallAngleSpinTiltCard()
-        card.canting_mode_combo.setCurrentText("Atom pair canting")
-        card.pair_source_combo.setCurrentText("Auto by neighbor shell")
-        card.pair_shell_frame.set_input_value([1])
-        card.pair_tol_frame.set_input_value([0.02])
-        card.pair_element_edit.setText("Fe-Co")
-        card.pair_group_edit.setText("A-B")
-        card.bond_mode_combo.setCurrentText("Near axis")
-        card.bond_axis_frame.set_input_value([0.0, 0.0, 1.0])
-        card.bond_tol_frame.set_input_value([5.0])
-        card.angle_edit.setText("6")
-        card.sign_combo.setCurrentText("Both (+/- pair)")
-        card.include_reference_checkbox.setChecked(False)
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        tags = [str(a.info.get("Config_type", "")) for a in results]
-        self.assertTrue(all("SpinPair(i=2,j=3,a=6" in tag for tag in tags))
-
-        data = card.to_dict()
-        restored = SmallAngleSpinTiltCard()
-        restored.from_dict(data)
-        self.assertEqual(restored.pair_source_combo.currentText(), "Auto by neighbor shell")
-        self.assertEqual(restored.pair_element_edit.text(), "Fe-Co")
-        self.assertEqual(restored.pair_group_edit.text(), "A-B")
-        self.assertEqual(restored.bond_mode_combo.currentText(), "Near axis")
-        self.assertEqual(restored.bond_axis_frame.get_input_value(), [0.0, 0.0, 1.0])
-        self.assertEqual(restored.bond_tol_frame.get_input_value(), [5.0])
-        self.assertFalse(restored.bond_axis_label.isHidden())
-        self.assertFalse(restored.bond_tol_label.isHidden())
-
-    def test_spin_disorder_operation_and_card_roundtrip(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([2.0, 2.0, 2.0, 2.0])
-
-        results = SpinDisorderOperation().run_structure(
-            structure,
-            SpinDisorderParams(
-                mode="Flip fraction",
-                fractions="0.5",
-                samples_per_fraction=1,
-                use_seed=True,
-                seed=11,
-            ),
-        )
-        self.assertEqual(len(results), 1)
-        moments = np.array(results[0].get_initial_magnetic_moments(), dtype=float)
-        self.assertEqual(int(np.sum(moments[:, 2] < 0.0)), 2)
-        self.assertIn("SpinDis(f=0.5,n=2,mode=flip", str(results[0].info.get("Config_type", "")))
-
-        cone = SpinDisorderOperation().run_structure(
-            structure,
-            SpinDisorderParams(
-                mode="Cone disorder",
-                fractions="0.5",
-                cone_angle=10.0,
-                use_seed=True,
-                seed=11,
-            ),
-        )[0]
-        cone_moments = np.array(cone.get_initial_magnetic_moments(), dtype=float)
-        self.assertTrue(np.allclose(np.linalg.norm(cone_moments, axis=1), 2.0, atol=1e-6))
-
-        card = SpinDisorderCard()
-        card.mode_combo.setCurrentText("Cone disorder")
-        card.fractions_edit.setText("0.25,0.5")
-        card.samples_frame.set_input_value([2])
-        card.cone_frame.set_input_value([12.0])
-        card.seed_checkbox.setChecked(True)
-        card.seed_frame.set_input_value([5])
-        restored = SpinDisorderCard()
-        restored.from_dict(card.to_dict())
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_set_magnetic_moments_card_map_vector_roundtrip(self):
-        structure = self._spin_chain()
-
-        card = SetMagneticMomentsCard()
-        card.source_combo.setCurrentText("Map/default magnitude")
-        card.format_combo.setCurrentText("Non-collinear (vector)")
-        card.map_edit.setText("Fe:2.5")
-        card.axis_frame.set_input_value([0.0, 0.0, 1.0])
-
-        result = card.process_structure(structure)[0]
-        moments = np.array(result.get_initial_magnetic_moments(), dtype=float)
-        self.assertEqual(moments.shape, (4, 3))
-        self.assertTrue(np.allclose(moments, np.tile([0.0, 0.0, 2.5], (4, 1)), atol=1e-6))
-        self.assertIn("MagSet(map,vec)", str(result.info.get("Config_type", "")))
-
-        data = card.to_dict()
-        restored = SetMagneticMomentsCard()
-        restored.from_dict(data)
-        self.assertEqual(restored.source_combo.currentText(), "Map/default magnitude")
-        self.assertEqual(restored.format_combo.currentText(), "Non-collinear (vector)")
-        self.assertEqual(restored.map_edit.text(), "Fe:2.5")
-
-    def test_set_magnetic_moments_card_existing_scalar_to_vector(self):
-        structure = self._spin_chain()
-        structure.set_initial_magnetic_moments([1.0, 2.0, 3.0, 4.0])
-
-        card = SetMagneticMomentsCard()
-        card.source_combo.setCurrentText("Existing initial magmoms")
-        card.format_combo.setCurrentText("Non-collinear (vector)")
-        card.axis_frame.set_input_value([1.0, 0.0, 0.0])
-
-        result = card.process_structure(structure)[0]
-        moments = np.array(result.get_initial_magnetic_moments(), dtype=float)
-        self.assertTrue(
-            np.allclose(
-                moments,
-                np.array([[1.0, 0.0, 0.0], [2.0, 0.0, 0.0], [3.0, 0.0, 0.0], [4.0, 0.0, 0.0]]),
-                atol=1e-6,
-            )
-        )
-        self.assertIn("MagSet(existing,vec)", str(result.info.get("Config_type", "")))
-
-    def test_magmom_rotation_lifts_scalar_to_vector(self):
-        proto = CrystalPrototypeBuilderCard()
-        proto.structure_combo.setCurrentText("bcc")
-        proto.element_edit.setText("Fe")
-        proto.a_frame.set_input_value([2.9, 2.9, 0.1])
-        proto.manual_supercell_button.setChecked(True)
-        proto.auto_supercell_button.setChecked(False)
-        proto.rep_frame.set_input_value([2, 2, 2])
-        proto.max_output_frame.set_input_value([1])
-        base = proto.create_operation().generate(proto.get_params())[0]
-
-        order = MagneticOrderCard()
-        order.format_combo.setCurrentText("Collinear (scalar)")
-        order.map_edit.setText("Fe:2.2")
-        order.fm_checkbox.setChecked(True)
-        order.afm_checkbox.setChecked(False)
-        order.pm_checkbox.setChecked(False)
-        fm = order.process_structure(base)[0]
-
-        rot = MagneticMomentRotationCard()
-        rot.elements_input.setText("Fe")
-        rot.angle_frame.set_input_value([45.0])
-        rot.lift_scalar_checkbox.setChecked(True)
-        rot.axis_frame.set_input_value([0.0, 0.0, 1.0])
-        rot.count_frame.set_input_value([1])
-
-        rotated = rot.process_structure(fm)[0]
-        m = np.array(rotated.get_initial_magnetic_moments(), dtype=float)
-        self.assertEqual(m.ndim, 2)
-        self.assertEqual(m.shape[1], 3)
-
-    def test_random_vacancy_card(self):
-        card = RandomVacancyCard()
-        structure = self.structure.copy()
-        card.rules_widget.from_rules([
-            {"element": "Si", "count": [1, 1], "count_mode": "fixed"},
-        ])
-        card.max_atoms_condition_frame.set_input_value([2])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        self.assertTrue(all(len(atoms) == len(structure) - 1 for atoms in results))
-
-    def test_vacancy_rule_count_mode_distinguishes_fixed_and_range(self):
-        item = VacancyRuleItem()
-        item.element_edit.setText("Si")
-        item.fixed_count_frame.set_input_value([2])
-        fixed = item.to_rule()
-        self.assertEqual(fixed["count"], [2, 2])
-        self.assertEqual(fixed["count_mode"], "fixed")
-
-        item.count_mode_combo.setCurrentText("Random range")
-        item.count_range_frame.set_input_value([1, 3])
-        ranged = item.to_rule()
-        self.assertEqual(ranged["count"], [1, 3])
-        self.assertEqual(ranged["count_mode"], "random")
-
-    def test_random_vacancy_operation_fixed_count(self):
-        structure = Atoms("Si5", positions=np.arange(15, dtype=float).reshape(5, 3), cell=[10, 10, 10], pbc=True)
-
-        results = RandomVacancyOperation().run_structure(
-            structure,
-            RandomVacancyParams(
-                rules=[{"element": "Si", "count": [2, 2], "count_mode": "fixed"}],
-                max_structures=4,
-                use_seed=True,
-                seed=1,
-            ),
-        )
-
-        self.assertTrue(all(len(atoms) == 3 for atoms in results))
-
-    def test_vacancy_defect_card_concentration(self):
-        card = VacancyDefectCard()
-        structure = self.structure.copy()
-        card.engine_type_combo.setCurrentIndex(1)
-        card.concentration_radio_button.setChecked(True)
-        card.num_radio_button.setChecked(False)
-        card.concentration_condition_frame.set_input_value([0.6])
-        card.num_condition_frame.set_input_value([1])
-        card.count_mode_combo.setCurrentText("Random up to value")
-        card.max_atoms_condition_frame.set_input_value([2])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        self.assertTrue(all(len(atoms) < len(structure) for atoms in results))
-
-    def test_vacancy_defect_fixed_count(self):
-        structure = Atoms("Si5", positions=np.arange(15, dtype=float).reshape(5, 3), cell=[10, 10, 10], pbc=True)
-
-        results = VacancyDefectOperation().run_structure(
-            structure,
-            VacancyDefectParams(
-                engine_type=1,
-                num_condition=3,
-                use_num=True,
-                count_mode="fixed",
-                max_structures=3,
-                use_seed=True,
-                seed=2,
-            ),
-        )
-
-        self.assertTrue(all(len(atoms) == 2 for atoms in results))
-
-    def test_stacking_fault_card(self):
-        card = StackingFaultCard()
-        structure = self.structure.copy()
-        card.hkl_frame.set_input_value([1, 1, 1])
-        card.layer_frame.set_input_value([1])
-        card.step_frame.set_input_value([0.1, 0.1, 0.1])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 1)
-        self.assertGreater(
-            np.abs(results[0].get_positions() - structure.get_positions()).sum(),
-            0.0,
-        )
-
-    def test_defect_surface_operations_are_ui_independent(self):
-        structure = self.structure.copy()
-
-        vac_results = RandomVacancyOperation().run_structure(
-            structure,
-            RandomVacancyParams(
-                rules=[{"element": "Si", "count": [1, 1], "count_mode": "fixed"}],
-                max_structures=2,
-                use_seed=True,
-                seed=3,
-            ),
-        )
-        self.assertEqual(len(vac_results), 2)
-        self.assertTrue(all(len(atoms) == len(structure) - 1 for atoms in vac_results))
-
-        defect_results = VacancyDefectOperation().run_structure(
-            structure,
-            VacancyDefectParams(
-                engine_type=1,
-                use_num=False,
-                concentration_condition=0.6,
-                count_mode="random",
-                max_structures=2,
-                use_seed=True,
-                seed=5,
-            ),
-        )
-        self.assertEqual(len(defect_results), 2)
-        self.assertTrue(all(len(atoms) < len(structure) for atoms in defect_results))
-
-        fault_results = StackingFaultOperation().run_structure(
-            structure,
-            StackingFaultParams(hkl=(1, 1, 1), step=(0.1, 0.1, 0.1), layers=1),
-        )
-        self.assertEqual(len(fault_results), 1)
-        self.assertIn("SF(", fault_results[0].info.get("Config_type", ""))
-
-        slab_results = RandomSlabOperation().run_structure(
-            structure,
-            RandomSlabParams(
-                h_range=(1, 1, 1),
-                k_range=(0, 0, 1),
-                l_range=(0, 0, 1),
-                layer_range=(1, 1, 1),
-                vacuum_range=(0, 0, 1),
-            ),
-        )
-        self.assertEqual(len(slab_results), 1)
-        self.assertIn("Slab(", slab_results[0].info.get("Config_type", ""))
-
-        insert_results = InsertDefectOperation().run_structure(
-            structure,
-            InsertDefectParams(
-                mode=0,
-                species="H",
-                insert_count=1,
-                structure_count=1,
-                min_distance=0.1,
-                max_attempts=20,
-                use_seed=True,
-                seed=7,
-            ),
-        )
-        self.assertEqual(len(insert_results), 1)
-        self.assertEqual(len(insert_results[0]), len(structure) + 1)
-        self.assertIn("Ins(int", insert_results[0].info.get("Config_type", ""))
-
-    def test_insert_defect_card_roundtrip(self):
-        card = InsertDefectCard()
-        card.mode_combo.setCurrentIndex(1)
-        card.species_edit.setText("H:2, O:1")
-        card.insert_count_frame.set_input_value([2])
-        card.structures_frame.set_input_value([3])
-        card.min_distance_frame.set_input_value([0.5])
-        card.max_attempts_frame.set_input_value([50])
-        card.seed_checkbox.setChecked(True)
-        card.seed_frame.set_input_value([13])
-        card.axis_combo.setCurrentIndex(2)
-        card.offset_frame.set_input_value([2.0])
-
-        restored = InsertDefectCard()
-        restored.from_dict(card.to_dict())
-
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_layer_copy_operation_is_ui_independent(self):
-        structure = self.structure.copy()
-        results = LayerCopyOperation().run_structure(
-            structure,
-            LayerCopyParams(
-                dz_expr="1.0",
-                layers=2,
-                distance=3.0,
-                extend_cell_z=True,
-            ),
-        )
-
-        self.assertEqual(len(results), 1)
-        self.assertEqual(len(results[0]), len(structure) * 2)
-        self.assertIn("SWC(L=2,dz=3)", results[0].info.get("Config_type", ""))
-
-    def test_layer_copy_card_roundtrip(self):
-        card = LayerCopyCard()
-        card.preset_combo.setCurrentIndex(0)
-        card.expr_edit.setPlainText("A + z*0")
-        card.params_edit.setText("A=1.5")
-        card.apply_combo.setCurrentIndex(2)
-        card.zrange_frame.set_input_value([0.0, 2.0])
-        card.wrap_checkbox.setChecked(True)
-        card.extend_cell_checkbox.setChecked(False)
-        card.vacuum_frame.set_input_value([1.0])
-        card.layers_frame.set_input_value([2])
-        card.distance_frame.set_input_value([4.0])
-
-        restored = LayerCopyCard()
-        restored.from_dict(card.to_dict())
-
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_organic_configuration_card(self):
-        card = OrganicMolConfigPBCCard()
-        structure = self.structure.copy()
-        card.perturb_frame.set_input_value([2])
-        card.torsion_frame.set_input_value([-30.0, 30.0])
-        card.max_torsions_frame.set_input_value([1])
-        card.sigma_frame.set_input_value([0.01])
-        card.pbc_combo.setCurrentIndex(1)
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 2)
-        for atoms in results:
-            self.assertEqual(len(atoms), len(structure))
-            self.assertFalse(np.allclose(atoms.get_positions(), structure.get_positions()))
-
-    def test_organic_configuration_card_roundtrip(self):
-        card = OrganicMolConfigPBCCard()
-        card.perturb_frame.set_input_value([3])
-        card.torsion_frame.set_input_value([-45.0, 60.0])
-        card.max_torsions_frame.set_input_value([2])
-        card.sigma_frame.set_input_value([0.02])
-        card.pbc_combo.setCurrentIndex(2)
-        card.local_cut_frame.set_input_value([80])
-        card.local_sub_frame.set_input_value([30])
-        card.bond_max_enable.setChecked(True)
-        card.seed_checkbox.setChecked(True)
-        card.seed_frame.set_input_value([11])
-
-        restored = OrganicMolConfigPBCCard()
-        restored.from_dict(card.to_dict())
-
-        self.assertEqual(restored.get_params(), card.get_params())
-
-    def test_vibration_mode_perturb_card(self):
-        card = VibrationModePerturbCard()
-        structure = self.structure.copy()
-        natoms = len(structure)
-        n_modes = min(3 * natoms, 6)
-        mode_vectors = np.zeros((n_modes, natoms, 3))
-        for idx in range(n_modes):
-            atom_index = idx % natoms
-            component = idx % 3
-            mode_vectors[idx, atom_index, component] = 1.0
-        freq_values = np.linspace(50.0, 300.0, n_modes)
-        for mode_idx in range(n_modes):
-            structure.new_array(f"vibration_mode_{mode_idx}_x", mode_vectors[mode_idx, :, 0])
-            structure.new_array(f"vibration_mode_{mode_idx}_y", mode_vectors[mode_idx, :, 1])
-            structure.new_array(f"vibration_mode_{mode_idx}_z", mode_vectors[mode_idx, :, 2])
-            structure.new_array(
-                f"vibration_frequency_{mode_idx}",
-                np.full(natoms, freq_values[mode_idx], dtype=float),
-            )
-
-        card.amplitude_frame.set_input_value([0.05])
-        card.modes_frame.set_input_value([2])
-        card.min_freq_frame.set_input_value([1.0])
-        card.num_condition_frame.set_input_value([3])
-
-        results = card.process_structure(structure)
-        self.assertEqual(len(results), 3)
-        base_positions = structure.get_positions()
-        displacements = [
-            np.linalg.norm(atoms.get_positions() - base_positions, axis=1).max()
-            for atoms in results
-        ]
-        self.assertTrue(all(delta > 0 for delta in displacements))
-        self.assertTrue(all(len(atoms) == len(structure) for atoms in results))
-
-        op_results = VibrationModePerturbOperation().run_structure(
-            structure,
-            VibrationModePerturbParams(
-                amplitude=0.05,
-                modes_per_sample=2,
-                min_frequency=1.0,
-                max_num=2,
-                use_seed=True,
-                seed=4,
-            ),
-        )
-        self.assertEqual(len(op_results), 2)
-
-        restored = VibrationModePerturbCard()
-        restored.from_dict(card.to_dict())
-        self.assertEqual(restored.get_params(), card.get_params())
-
-
-if __name__ == "__main__":
-    unittest.main()
